Fixed issue # 1 Application that was run with the BTrace agent hangs, need to use CTRL-C to exit.
A thread pool is created to get low memory alerts. The thread factory to
create threads for this pool has to create daemon threads. Without daemon
threads, the application waits for these threads to finish. With the change
to create daemon threads, user need not have to CRTL-C to exit.

Fixed issue # 1 Application that was run with the BTrace agent hangs, need to use CTRL-C to exit.

A thread pool is created to get low memory alerts. The thread factory to
create threads for this pool has to create daemon threads. Without daemon
threads, the application waits for these threads to finish. With the change
to create daemon threads, user need not have to CRTL-C to exit.