When tracing JavaFX apps, I was not able to put probes in a particular class (com.sun.javafx.runtime.FXBase). It turned out that the problem is about timing. BTrace gets all loaded classes and filters the list to create candidate list. But, classes may be loaded after "getAllLoadedClasses" call, but before the bytecode transformer is added (which is added after filtering is completed).
So, I moved "inst.addTransformer(client, true)" so that "client" won't miss
the classes?loaded during the analysis of snapshot of all-loaded-classes.
With this change, I was able to put probed in FXBase for the JavaFX app.

Testing: I was not able to run tests properly. But, tests fail even without
my fix. So, I am going ahead with the fix.

    [junit] Testcase: bytecodeValidation(com.sun.btrace.runtime.BTRACE22Test):	Caused an ERROR
    [junit] null
    [junit] java.lang.NullPointerException
    [junit] 	at support.InstrumentorTestBase.loadFile(InstrumentorTestBase.java:165)
    [junit] 	at support.InstrumentorTestBase.loadTrace(InstrumentorTestBase.java:149)
    [....]